fix(tui): u16 overflow in sidebar width, zero-width char in word break
- settings overlay: widen to u32 before multiplying area.width * 35
  to avoid u16 overflow on terminals wider than 1875 columns
- auto_drive celebration: use unwrap_or(0) for control/combining
  chars in word-break width calculation (consistency with other sites)

+++ code-rs/tui/src/chatwidget/settings_overlay/overlay_render/section.rs
@@ -40,7 +40,7 @@ impl SettingsOverlayView {
             // Expanded: sidebar (with toggle row on top) + content.
             // Adaptive width: cap at 22 but shrink proportionally on
             // narrow screens so the content pane always gets ≥ 60% width.
-            let sidebar_width = 22u16.min(area.width * 35 / 100).max(12);
+            let sidebar_width = 22u16.min((u32::from(area.width) * 35 / 100) as u16).max(12);
             let [sidebar_col, main] =
                 Layout::horizontal([Constraint::Length(sidebar_width), Constraint::Fill(1)])
                     .areas(area);
